Port to 25w05a, removed all dependency and plugin

+++ src/main/java/com/chesy/productiveslimes/block/custom/SlimeBlock.java
@@ -34,7 +34,7 @@ public class SlimeBlock extends TranslucentBlock {
     }
 
     @Override
-    public void onLandedUpon(World world, BlockState state, BlockPos pos, Entity entity, float fallDistance) {
+    public void onLandedUpon(World world, BlockState state, BlockPos pos, Entity entity, double fallDistance) {
         if (!entity.bypassesLandingEffects()) {
             entity.handleFallDamage(fallDistance, 0.0F, world.getDamageSources().fall());
         }
